Some api work
Now on to client side logging

+++ Discord Groups/Discord Groups - Server/src/com/weeryan17/dgs/listeners/discord/ChatListener.java
@@ -23,22 +23,24 @@ public class ChatListener {
 	public void onMessage(MessageReceivedEvent e){
 		IMessage message = e.getMessage();
 		String text = message.getContent();
-		if(prefix == text.charAt(0)){
-			text = text.substring(1);
-			String[] command = text.split(" ");
-			String name = command[0];
-			String[] args = Arrays.copyOfRange(command, 1, command.length);
-			if(instance.getCommandMannager().isCommand(name)){
-				message.delete();
-				instance.getCommandMannager().dispatchCommand(name, args, e.getChannel(), e.getAuthor());
-				String compressedArgs = "";
-				for(String arg : args){
-					compressedArgs = compressedArgs + "\n" + space + " * " + arg;
+		if(text.length() >= 1){
+			if(prefix == text.charAt(0)){
+				text = text.substring(1);
+				String[] command = text.split(" ");
+				String name = command[0];
+				String[] args = Arrays.copyOfRange(command, 1, command.length);
+				if(instance.getCommandMannager().isCommand(name)){
+					message.delete();
+					instance.getCommandMannager().dispatchCommand(name, args, e.getChannel(), e.getAuthor());
+					String compressedArgs = "";
+					for(String arg : args){
+						compressedArgs = compressedArgs + "\n" + space + " * " + arg;
+					}
+					instance.getLogger().log("The user " + e.getAuthor().getName() + " who's id is " + e.getAuthor().getID() + "\n" + 
+						space + "executed the command: " + name + "\n" + space + "With the args:" + compressedArgs, false);
+				} else {
+					message.addReaction(EmojiManager.getForAlias("question"));
 				}
-				instance.getLogger().log("The user " + e.getAuthor().getName() + " who's id is " + e.getAuthor().getID() + "\n" + 
-					space + "executed the command: " + name + "\n" + space + "With the args:" + compressedArgs, false);
-			} else {
-				message.addReaction(EmojiManager.getForAlias("question"));
 			}
 		}
 	}
